* Merge of trunk with /branches/arl_dev     (svn merge -r53819:53890)
* Generate new configure script


Former-commit-id: eb66b10b90d7acf0dcdffb9efbdf7242a6692c21

+++ src/CCA/Components/MPM/ParticleCreator/ImplicitParticleCreator.h
@@ -44,7 +44,7 @@ namespace Uintah {
                                     MPMMaterial* matl,
                                     Point p, IntVector cell_idx,
                                     particleIndex i,
-                                    CCVariable<short int>& cellNAPI,
+                                    CCVariable<int>& cellNAPI,
                                     ParticleVars& pvars);
 
     virtual ParticleSubset* allocateVariables(particleIndex numParticles,
